Moved all tests regarding debug naming to the new TU debug_naming_tests
